dev-util/jd-gui: fix install path

+++ dev-util/jd-gui/jd-gui-1.6.6.ebuild
@@ -25,16 +25,22 @@ S="${WORKDIR}/opt/jd-gui"
 src_install() {
 	insinto /opt/"${MY_PN}"
 	dodir /opt/"${MY_PN}"
-	mv jd_*.jar "${MY_P}.jar" || die
+	if [ ! -f "${MY_P}.jar" ]; then
+		mv jd*.jar "${MY_P}.jar" || die
+	fi
 	doins "${MY_P}.jar"
 
 	echo -e "#!/bin/sh\njava -jar /opt/${MY_PN}/${MY_P}.jar >/dev/null 2>&1 &\n" > "${MY_PN}"
 	dobin "${MY_PN}"
 
-	mv jd_*.png "${MY_PN}.png" || die
+	if [ ! -f "${MY_PN}.png" ]; then
+		mv jd*.png "${MY_PN}.png" || die
+	fi
 	doicon -s 128 "${MY_PN}.png"
 
-	mv jd_*.desktop "${MY_PN}.desktop" || die
+	if [ ! -f "${MY_PN}".desktop ]; then
+		mv jd*.desktop "${MY_PN}.desktop" || die
+	fi
 	sed -i "s|Exec=.*$|Exec=java -jar /opt/${MY_PN}/${MY_P}.jar|g" "${MY_PN}".desktop
 	sed -i "s|Icon=.*$|Icon=${MY_PN}|g" "${MY_PN}".desktop
 	domenu "${MY_PN}".desktop
